feat(migration) : Passage à Java 11 et Spring Boot 2.1
- Migration des versions de Java et Spring Boot
- Properties to YAML
- Changement de packages
