--- src/main/java/com/cpetot/exceptions/UserNameNotFoundException.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2012-2016 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.cpetot.exceptions;
-
-public class UserNameNotFoundException extends RuntimeException {
-
-	private static final long serialVersionUID = 1L;
-	private String username;
-
-	public UserNameNotFoundException(String username) {
-		super(String.format("User with username %s does not exist", username));
-		this.username = username;
-	}
-
-	public String getUsername() {
-		return this.username;
-	}
-
-}
