--- src/main/java/com/cpetot/repository/MovieRepository.java
@@ -1,11 +1,0 @@
-package com.cpetot.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.cpetot.entities.Movie;
-
-public interface MovieRepository extends JpaRepository<Movie, Long> {
-
-	Movie findByTitle(String title);
-
-}
