--- src/main/java/com/cpetot/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.cpetot.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.cpetot.entities.User;
-
-public interface UserRepository extends JpaRepository<User, Long>{
-
-	User findByUsername(String username);
-
-}
