--- src/main/java/com/cpetot/web/RestResponse.java
@@ -1,26 +1,0 @@
-package com.cpetot.web;
-
-public class RestResponse {
-
-	private boolean success;
-	private String message;
-
-	public RestResponse(boolean success, String message) {
-		this.success = success;
-		this.message = message;
-	}
-
-	public boolean getSuccess() {
-		return success;
-	}
-	public void setSuccess(boolean success) {
-		this.success = success;
-	}
-	public String getMessage() {
-		return message;
-	}
-	public void setMessage(String message) {
-		this.message = message;
-	}
-
-}
