+++ src/main/java/com/cpetot/workshop/tu/WorkshopUnitTestsApplication.java
@@ -1,0 +1,14 @@
+package com.cpetot.workshop.tu;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Import;
+
+@SpringBootApplication
+@Import(SwaggerWebConfiguration.class)
+public class WorkshopUnitTestsApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(WorkshopUnitTestsApplication.class, args);
+	}
+}
