--- src/main/resources/application.properties
@@ -1,13 +1,0 @@
-# Database configuration
-spring.datasource.url=jdbc\:h2\:./development.h2.db
-spring.datasource.username=sa
-spring.datasource.password=
-spring.datasource.driver-class-name=org.h2.Driver
-spring.datasource.platform=h2
-
-# H2
-spring.h2.console.enabled=true
-
-# Hibernate
-logging.level.org.hibernate.SQL=DEBUG
-logging.level.org.hibernate.type.descriptor.sql=TRACE
