--- src/test/java/com/cpetot/service/MovieServiceTest.java
@@ -1,8 +1,0 @@
-package com.cpetot.service;
-
-public class MovieServiceTest {
-
-	// Tests de la méthode getByTitle
-
-	// Tests de la méthode findMoviesAvailableForAge
-}
