+++ src/test/java/com/cpetot/workshop/tu/service/MovieServiceTest.java
@@ -1,0 +1,8 @@
+package com.cpetot.workshop.tu.service;
+
+public class MovieServiceTest {
+
+	// Tests de la méthode getByTitle
+
+	// Tests de la méthode findMoviesAvailableForAge
+}
